+++ src/idunno.AtProto/SourceGenerationContext.cs
@@ -42,6 +42,7 @@ namespace idunno.AtProto
     [JsonSerializable(typeof(Handle))]
     [JsonSerializable(typeof(Nsid))]
     [JsonSerializable(typeof(RecordKey))]
+    [JsonSerializable(typeof(TimestampIdentifier))]
 
     [JsonSerializable(typeof(SubjectType))]
     [JsonSerializable(typeof(StrongReference))]