+++ test/idunno.AtProto.Types.Test/TimestampIdentifierTests.cs
@@ -1,6 +1,11 @@
 ﻿// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Runtime.CompilerServices;
+using System.Security.Cryptography;
+using System.Text.Json.Serialization;
+using System.Xml.Serialization;
+
 namespace idunno.AtProto.Types.Test
 {
     public class TimestampIdentifierTests
@@ -16,7 +21,7 @@ namespace idunno.AtProto.Types.Test
         }
 
         [Fact]
-        public void TimestampIdentifierMeetsSpec()
+        public void TimestampIdentifierNextMeetsSpec()
         {
             RecordKey now = TimestampIdentifier.Next();
 
@@ -88,5 +93,101 @@ namespace idunno.AtProto.Types.Test
             Assert.Equal(timeStamp, tid.TimeStamp);
             Assert.Equal(clockId, tid.ClockId);
         }
+
+        [Fact]
+        public void TimestampIdentifierMeetsSpec()
+        {
+            var actual = new TimestampIdentifier();
+
+            Assert.Matches("""^[234567abcdefghij][234567abcdefghijklmnopqrstuvwxyz]{12}$""", actual.ToString());
+        }
+
+        [Fact]
+        public void ImplicitConversionToRecordKeyWorks()
+        {
+            var tid = new TimestampIdentifier();
+            RecordKey rk = (RecordKey)tid;
+            Assert.Equal(tid.ToString(), rk.ToString());
+        }
+
+        [Fact]
+        public void ExplicitConversionToRecordKeyWorks()
+        {
+            var tid = new TimestampIdentifier();
+            RecordKey rk = tid.ToRecordKey();
+            Assert.Equal(tid.ToString(), rk.ToString());
+        }
+
+        [Fact]
+        public void TimestampIdentifierConstructorThrowsOnNullString()
+        {
+            Assert.Throws<ArgumentNullException>(() => new TimestampIdentifier(null!));
+        }
+
+        [Fact]
+        public void TimestampIdentifierConstructorThrowsOnInvalidLengthString()
+        {
+            Assert.Throws<ArgumentException>(() => new TimestampIdentifier("short"));
+        }
+
+        [Fact]
+        public void TimestampIdentifierConstructorThrowsOnInvalidFormatString()
+        {
+            Assert.Throws<ArgumentException>(() => new TimestampIdentifier("invalidFormat"));
+        }
+
+        [Fact]
+        public void TimestampIdentifierConstructorFromRecordKeyWorks()
+        {
+            var expected = new RecordKey("3lwjaurx2d22c");
+
+            var actual = new TimestampIdentifier(expected);
+            Assert.Equal(expected.ToString(), actual.ToString());
+        }
+
+        // Values taken from https://atproto.com/specs/tid#examples
+        [Theory]
+        [InlineData("3jzfcijpj2z2a")]
+        [InlineData("7777777777777")]
+        [InlineData("3zzzzzzzzzzzz")]
+        [InlineData("2222222222222")]
+        public void SyntacticallyValidTimestampIdentifiersAreAccepted(string tidString)
+        {
+            var tid = new TimestampIdentifier(tidString);
+            Assert.Equal(tidString, tid.ToString());
+        }
+
+        [Theory]
+        [InlineData("3jzfcijpj2z21")]
+        [InlineData("0000000000000")]
+        [InlineData("3JZFCIJPJ2Z2A")]
+        [InlineData("3jzfcijpj2z2aa")]
+        [InlineData("3jzfcijpj2z2")]
+        [InlineData("222")]
+        [InlineData("zzzzzzzzzzzzz")]
+        [InlineData("kjzfcijpj2z2a")]
+        public void InvalidTimestampIdentifiersCannotBeConstructed(string tidAsString)
+        {
+            Assert.Throws<ArgumentException>(() => _ = new TimestampIdentifier(tidAsString));
+        }
+
+        [Fact]
+        public void TimestampIdentifierDeserializationWorks()
+        {
+            string json = """{"tid":"3lwjaurx2d22c"}""";
+
+            var expected = new TimestampIdentifier("3lwjaurx2d22c");
+
+            TimestampIdentifierRecord? actual = System.Text.Json.JsonSerializer.Deserialize<TimestampIdentifierRecord>(json, System.Text.Json.JsonSerializerOptions.Web);
+
+            Assert.Equal(actual!.Tid.Value, expected.Value);
+        }
+
+    }
+
+    internal record TimestampIdentifierRecord
+    {
+        [JsonRequired]
+        public required TimestampIdentifier Tid { get; init; }
     }
 }